Use plugin for top-level resolution per entry
We had been including each entry directory as a resolution point for
webpack so that we could require modules from the top-level entry
directories down. Unfortunately that meant same-named sub-directories
couldn’t be differentiated and so they’d conflict.

Using the `resolve-entry-modules-webpack-plugin` means each entry is
quarantined from the other.